requirements.txt
imutils==0.5.3
tensorflow==2.0.0
numpy==1.17.4
streamlit==0.73.1
Pillow==8.1.0
